dhcpv6-ia: add option to roll DHCPv6 IA_NA
FS#403. DHCPv6 has the potential for external network snooping and
enumeration, because HOSTID are so unique. Furthermore, if similar
hardware and software are used, then HOSTID from hashed DUID would
be same on multiple networks. A device could be traceable when we
should not expect it to be so obviously located.

Add `option dhcpv6_hashtime s/m/h/d/w` to occassionally change the
seed used to hash the DUID into HOSTID. A common use case would be
for 12 or 24 hours so that all HOSTID are new at least each day.
This is intended for obscurity on IA_NA "non-temporary" HOSTID,
and this is not a replacement for greater privacy within IA_TA.

Signed-off-by: Eric Luehrsen <ericluehrsen@hotmail.com>